+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added support for a nonstrict NPCSpawns directory
 * Added support for the DisplayAction attribute
 * Added a lint for a Geometry folder
+* Updated the ButtonInteraction format
+* Removed Cases
 
 ## [v0.7.0] - August 8, 2023
 * Added a lint group for Conditionals